requirements.txt
alembic==1.10.2
bcrypt==4.0.1
Bcrypt-Flask==1.0.2
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
python-dotenv==1.0.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
